requirements.txt
yfinance
python-telegram-bot
pandas
openpyxl
requests